Refactored image selection into UIImagePicker extension.
