+++ Seattle Trails/AppExtensions.swift
@@ -13,11 +13,43 @@ extension UIImagePickerController {
     func getPictureFor<VC: UIViewController where VC: UIImagePickerControllerDelegate>(purpose: String, sender: VC) {
         if UIImagePickerController.isSourceTypeAvailable(UIImagePickerControllerSourceType.Camera)
         {
-            
+            self.presentImageSourceSelectionView(sender: sender)
         }
         else
         {
             self.presentImagePickerWithSourceTypeForViewController(sender, sourceType: UIImagePickerControllerSourceType.PhotoLibrary)
+        }
+    }
+    
+    func presentImageSourceSelectionView<VC: UIViewController where VC: UIImagePickerControllerDelegate> (sender sender: VC) {
+        // Present image picker options.
+        let actionSheet = UIAlertController(title: "Image Source", message: nil, preferredStyle: UIAlertControllerStyle.ActionSheet)
+        
+        let cameraAction = UIAlertAction(title: "Camera", style: UIAlertActionStyle.Default)
+        { (action) in
+            dispatch_async(dispatch_get_main_queue())
+            {
+                self.presentImagePickerWithSourceTypeForViewController(sender, sourceType: UIImagePickerControllerSourceType.Camera)
+            }
+        }
+        
+        let libraryAction = UIAlertAction(title: "Photo Library", style: UIAlertActionStyle.Default)
+        { (action) in
+            dispatch_async(dispatch_get_main_queue())
+            {
+                self.presentImagePickerWithSourceTypeForViewController(sender, sourceType: UIImagePickerControllerSourceType.PhotoLibrary)
+            }
+        }
+        
+        let cancelAction = UIAlertAction(title: "Cancel", style: .Cancel, handler: nil)
+        
+        actionSheet.addAction(cameraAction)
+        actionSheet.addAction(libraryAction)
+        actionSheet.addAction(cancelAction)
+        
+        dispatch_async(dispatch_get_main_queue())
+        {
+            sender.presentViewController(actionSheet, animated: true, completion: nil)
         }
     }
     
@@ -32,12 +64,8 @@ extension UIImagePickerController {
 }
 
 extension UIImagePickerControllerDelegate {
-    var imagePicker: UIImagePickerController?
+    var imagePicker: UIImagePickerController!
     {
-        guard let _ = self.imagePicker else {
-            return nil
-        }
-        
         return UIImagePickerController()
     }
 }
